iio: adc: adrv9002: fix tdd rate calculation
The TDD core rate is calculated by dds_rate * rx_rate. In
commit 3fc394629138 ("iio: adc: adrv9002: fix tdd clock export") the
logic was changed to fix another bug so that the current channel when
dealing with TDD rate is a TX port. Hence 'chann->rate' cannot be used
as we want the RX rate which is not guaranteed to be the same as TX.

Fixes: 3fc394629138 ("iio: adc: adrv9002: fix tdd clock export")
Signed-off-by: Nuno Sá <nuno.sa@analog.com>
(cherry picked from commit b59cab3fcd1b07dcc58d248b16cb3b4a4b0603ee)